requirements.txt
Flask==3.0.3
dnspython==2.4.2
python-dotenv==1.0.1 